\name{Kardashian-package}
\alias{Kardashian-package}
\alias{Kardashian}
\docType{package}
\title{
Calculates the K-index by Hall (2014)
}
\description{
Seriously, all it does is calculate the Kardashian index when you input the number of Twitter followers and citations for a scientist. More info: http://genomebiology.com/2014/15/7/424
}
\details{
\tabular{ll}{
Package: \tab Kardashian\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-29\cr
License: \tab GPL-2\cr
}}
\author{
Andrew Kniss

Maintainer: Yeah... it probably won't be maintained... <akniss@uwyo.edu>
}
\references{
Hall, N. (2014) The Kardashian index: a measure of discrepant social media profile for scientists. Genome Biology. 15:24. doi:10.1186/s13059-014-0424-0 http://genomebiology.com/2014/15/7/424}
\keyword{ package }
\seealso{
}
\examples{
## For a scientist with 1000 Twitter followers and 10 citations:
Kindex(1000, 10)}
